+++ src/types/boolobject.rs
@@ -2,11 +2,11 @@
 use crate::ffi;
 use crate::object::PyObject;
 use crate::types::PyAny;
-use crate::AsPyPointer;
 use crate::FromPyObject;
 use crate::PyResult;
 use crate::Python;
-use crate::{IntoPyObject, PyTryFrom, ToPyObject};
+use crate::{AsPyPointer, FromPy};
+use crate::{PyTryFrom, ToPyObject};
 
 /// Represents a Python `bool`.
 #[repr(transparent)]
@@ -45,10 +45,10 @@ impl ToPyObject for bool {
     }
 }
 
-impl IntoPyObject for bool {
+impl FromPy<bool> for PyObject {
     #[inline]
-    fn into_object(self, py: Python) -> PyObject {
-        PyBool::new(py, self).into()
+    fn from_py(other: bool, py: Python) -> Self {
+        PyBool::new(py, other).into()
     }
 }