+++ src/types/floatob.rs
@@ -8,11 +8,11 @@ use crate::instance::PyNativeType;
 use crate::object::PyObject;
 use crate::objectprotocol::ObjectProtocol;
 use crate::types::PyAny;
-use crate::AsPyPointer;
 use crate::FromPyObject;
 use crate::PyResult;
 use crate::Python;
-use crate::{IntoPyObject, ToPyObject};
+use crate::ToPyObject;
+use crate::{AsPyPointer, FromPy};
 use std::os::raw::c_double;
 
 /// Represents a Python `float` object.
@@ -44,9 +44,9 @@ impl ToPyObject for f64 {
     }
 }
 
-impl IntoPyObject for f64 {
-    fn into_object(self, py: Python) -> PyObject {
-        PyFloat::new(py, self).into()
+impl FromPy<f64> for PyObject {
+    fn from_py(other: f64, py: Python) -> Self {
+        PyFloat::new(py, other).into()
     }
 }
 
@@ -70,9 +70,9 @@ impl ToPyObject for f32 {
     }
 }
 
-impl IntoPyObject for f32 {
-    fn into_object(self, py: Python) -> PyObject {
-        PyFloat::new(py, f64::from(self)).into()
+impl FromPy<f32> for PyObject {
+    fn from_py(other: f32, py: Python) -> Self {
+        PyFloat::new(py, f64::from(other)).into()
     }
 }