+++ CHANGELOG.md
@@ -5,7 +5,7 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
-## [Unreleased]
+## [0.8.0]
 
 ### Added
 
@@ -13,10 +13,11 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
  * `py_run!` macro [#512](https://github.com/PyO3/pyo3/pull/512)
  * Use existing fields and methods before calling custom __getattr__. [#505](https://github.com/PyO3/pyo3/pull/512)
  * `PyBytes` can now be indexed just like `Vec<u8>`
- * Implement `IntoPyObject` for `PyRef` and `PyRefMut`.
+ * Implement `IntoPy<PyObject>` for `PyRef` and `PyRefMut`.
 
 ## Removed
 
+ * `IntoPyObject` was replaced with `IntoPy<PyObject>`
  * `#[pyclass(subclass)]` is hidden a `unsound-subclass` feature because it's causing segmentation faults.
 
 ### Fixed